live chat app/Chat-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
eventlet==0.33.3       # for SocketIO server (or gevent)
